TST: Cleanup binding tests with conftest.py
